+++ src/pl-global.h
@@ -398,7 +398,7 @@ typedef struct PL_local_data
 
   struct
   { char *	_CWDdir;
-    int		_CWDlen;
+    size_t	_CWDlen;
 #ifdef __BEOS__
     status_t	dl_error;		/* dlopen() emulation in pl-beos.c */
 #endif
